Reorganize URI::Redis and URI::Valkey
