+++ .rubocop.yml
@@ -12,6 +12,11 @@ Metrics/ModuleLength:
   Exclude:
     - "spec/**/*"
 
+Naming/FileName:
+  Exclude:
+    - lib/uri-redis.rb
+    - lib/uri-valkey.rb
+
 RSpec/BeforeAfterAll:
   Enabled: false
 Lint/Void: # new in 1.1
@@ -25,15 +30,12 @@ Metrics/PerceivedComplexity: # new in 1.0
 Style/FormatStringToken: # new in 1.1
   Enabled: false
 Style/StringLiterals:
-  Enabled: true
   EnforcedStyle: single_quotes
 Style/StringLiteralsInInterpolation:
-  Enabled: true
   EnforcedStyle: single_quotes
 Style/FrozenStringLiteralComment:
   Enabled: false
 Layout/LineLength:
   Max: 120
 RSpec/ExampleLength:
-  Enabled: true
   Max: 10
