+++ lib/uri-valkey.rb
@@ -2,4 +2,4 @@
 
 # Compatibility shim for uri-valkey.rb
 # This file maintains backward compatibility for code that requires 'uri-valkey'
-require_relative 'uri_valkey'
+require_relative 'uri/valkey'
