+++ lib/uri_redis.rb
@@ -1,1 +1,5 @@
-require_relative 'uri_redis/redis'
+# frozen_string_literal: true
+
+# Compatibility shim for uri_redis.rb
+# This file maintains backward compatibility for code that requires 'uri_redis'
+require_relative 'uri/redis'
