+++ lib/uri_valkey.rb
@@ -1,1 +1,5 @@
-require_relative 'uri_valkey/valkey'
+# frozen_string_literal: true
+
+# Compatibility shim for uri_valkey.rb
+# This file maintains backward compatibility for code that requires 'uri_valkey'
+require_relative 'uri/valkey'
